+++ npx/createTemplate.js
@@ -35,6 +35,7 @@ import { getEventListenerTemp } from "./helpers/eventListeners.js";
 import { getFileTemp } from "./helpers/file.js";
 import { getPostMessageTemp } from "./helpers/postMessage.js";
 import { getArticleTemp } from "./helpers/articleSlot.js";
+import { getArticleNavigationTemp } from "./helpers/articleNavigation.js";
 
 const CURR_DIR = process.cwd();
 
@@ -130,6 +131,7 @@ export default plugin;
         const withFile = scopes.includes("File");
         const withPostMessage = scopes.includes("PostMessage");
         const withArticle = scopes.includes("ArticleButton");
+        const withArticleNavigation = scopes.includes("ArticleNavigation");
 
         const { apiVars, apiMeth, IApiPlugin } = getApiTemp(withApi);
         const { settingsVars, settingsMeth, ISettingsPlugin, ISettings } =
@@ -178,6 +180,12 @@ export default plugin;
           articleButtonVars,
           articleButtonMeth,
         } = getArticleTemp(withArticle);
+        const {
+          IArticleNavigationPlugin,
+          IArticleNavigationItem,
+          articleNavigationVars,
+          articleNavigationMeth,
+        } = getArticleNavigationTemp(withArticleNavigation);
 
         if (withApi) {
           pluginsImpIns += `, ${IApiPlugin}`;
@@ -229,6 +237,11 @@ export default plugin;
           pluginsIns += `, ${IArticleButtonPlugin}`;
         }
 
+        if (withArticleNavigation) {
+          pluginsImpIns += `, ${IArticleNavigationPlugin}, ${IArticleNavigationItem} `;
+          pluginsIns += `, ${IArticleNavigationPlugin}`;
+        }
+
         let nameIns = `${pluginName}`;
         let contentIns = `
   ${status}
@@ -242,6 +255,7 @@ export default plugin;
           ${fileVars}
           ${postMessageVars}
           ${articleButtonVars}
+          ${articleNavigationVars}
           ${onLoadCallback}
           ${updateStatus}
           ${getStatus}
@@ -255,7 +269,8 @@ export default plugin;
           ${eventListenerMeth}
           ${fileMeth}
           ${postMessageMeth}
-          ${articleButtonMeth}`;
+          ${articleButtonMeth}
+          ${articleNavigationMeth}`;
 
         template = template
           .replaceAll("pluginsImpIns", pluginsImpIns)