+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 3.0.0
+
+## Added
+
+- Add IArticleNavigationPlugin, IArticleNavigationItem, Section enum
+- Add `Actions.updateArticleNavigationItems` to refresh the article navigation
+  items and the open plugin section
+
 ## 2.1.0
 
 ## Changed